Moving the google calendar function into tools.py
